+++ wasm-node/CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Fixed
 
+- Fix finality stalling on epoch change by explicitly requesting justifications of blocks that a peer has reported as finalized but that isn't finalized locally. ([#441](https://github.com/smol-dot/smoldot/pull/441))
 - Fix `AlreadyDestroyedError` not being properly thrown if a function is called after `terminate()`. ([#438](https://github.com/smol-dot/smoldot/pull/438))
 
 ## 1.0.2 - 2023-04-12